ぶっ飛んだりするようなった
ただしアイテムを持ったりするキー配置も一緒になってるからはよ修正
