+++ Assets/script/ObjectRotation.cs
@@ -8,6 +8,9 @@ public class ObjectRotation : MonoBehaviour
     [SerializeField] Vector3 axis;
     public float rot;
     public float rotSpeed;
+    public bool up;
+    public bool right;
+    public bool forward;
     // Start is called before the first frame update
     void Start()
     {
@@ -17,10 +20,22 @@ public class ObjectRotation : MonoBehaviour
     // Update is called once per frame
     void Update()
     {
-        rot += rotSpeed*Time.deltaTime;
+        //rot += rotSpeed*Time.deltaTime;
         if (obj != null)
         {
-            obj.transform.Rotate(Vector3.right, rot);
+            if(up)
+            {
+                obj.transform.Rotate(Vector3.up, rot);
+            }
+            if (right)
+            {
+                obj.transform.Rotate(Vector3.right, rot);
+            }
+            if(forward)
+            {
+                obj.transform.Rotate(Vector3.forward, rot);
+            }
+
         }
     }
 }
