+++ Assets/script/P1InputControll.cs
@@ -1,0 +1,19 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class P1InputControll : MonoBehaviour
+{
+    [SerializeField] CharaCon cc;
+    // Start is called before the first frame update
+    void Start()
+    {
+
+    }
+
+    // Update is called once per frame
+    void Update()
+    {
+        cc.SetDirection(Input.GetAxis("Vertical"), Input.GetAxis("Horizontal"));
+    }
+}
