+++ Assets/script/P1InputControll.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: be0d8137cab60de40b4242b9cd8cbb6a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
