+++ Assets/script/P2InputControll.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2caa1ebb515bbd54c8b19907fa52e1af
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
