--- Assets/script/move2.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 030499a4bdb6a6f4ca39844733b72765
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
